move `host_page_size` to sysdeps
