+++ lucet-runtime/lucet-runtime-internals/src/alloc/mod.rs
@@ -1,37 +1,11 @@
 use crate::error::Error;
 use crate::module::Module;
 use crate::region::RegionInternal;
+use crate::sysdeps::host_page_size;
 use libc::c_void;
 use lucet_module::GlobalValue;
-use nix::unistd::{sysconf, SysconfVar};
 use rand::{thread_rng, Rng, RngCore};
-use std::sync::{Arc, Mutex, Once, Weak};
-
-pub const HOST_PAGE_SIZE_EXPECTED: usize = 4096;
-static mut HOST_PAGE_SIZE: usize = 0;
-static HOST_PAGE_SIZE_INIT: Once = Once::new();
-
-/// Our host is Linux x86_64, which should always use a 4K page.
-///
-/// We double check the expected value using `sysconf` at runtime.
-pub fn host_page_size() -> usize {
-    unsafe {
-        HOST_PAGE_SIZE_INIT.call_once(|| match sysconf(SysconfVar::PAGE_SIZE) {
-            Ok(Some(sz)) => {
-                if sz as usize == HOST_PAGE_SIZE_EXPECTED {
-                    HOST_PAGE_SIZE = HOST_PAGE_SIZE_EXPECTED;
-                } else {
-                    panic!(
-                        "host page size was {}; expected {}",
-                        sz, HOST_PAGE_SIZE_EXPECTED
-                    );
-                }
-            }
-            _ => panic!("could not get host page size from sysconf"),
-        });
-        HOST_PAGE_SIZE
-    }
-}
+use std::sync::{Arc, Mutex, Weak};
 
 pub fn instance_heap_offset() -> usize {
     1 * host_page_size()
