+++ lucet-runtime/lucet-runtime-internals/src/region/mmap.rs
@@ -1,9 +1,10 @@
-use crate::alloc::{host_page_size, instance_heap_offset, Alloc, AllocStrategy, Limits, Slot};
+use crate::alloc::{instance_heap_offset, Alloc, AllocStrategy, Limits, Slot};
 use crate::embed_ctx::CtxMap;
 use crate::error::Error;
 use crate::instance::{new_instance_handle, Instance, InstanceHandle};
 use crate::module::Module;
 use crate::region::{Region, RegionCreate, RegionInternal};
+use crate::sysdeps::host_page_size;
 use libc::c_void;
 #[cfg(not(target_os = "linux"))]
 use libc::memset;
