+++ lucet-runtime/lucet-runtime-internals/src/sysdeps/mod.rs
@@ -4,8 +4,14 @@ mod macos;
 #[cfg(target_os = "linux")]
 mod linux;
 
+#[cfg(unix)]
+mod unix;
+
 #[cfg(target_os = "macos")]
 pub use macos::*;
 
 #[cfg(target_os = "linux")]
 pub use linux::*;
+
+#[cfg(unix)]
+pub use unix::*;
